+++ packages/theme/package.json
@@ -42,6 +42,7 @@
     "react-dom": "16.8.6"
   },
   "dependencies": {
+    "@microsoft/load-themed-styles": "^1.10.26",
     "@uifabric/merge-styles": "^7.19.1",
     "@uifabric/utilities": "^7.32.3",
     "@uifabric/set-version": "^7.0.23",
