+++ packages/theme/src/utilities/makeSemanticColors.ts
@@ -1,0 +1,167 @@
+import { IEffects, IPalette, ISemanticColors } from '../types/index';
+
+/** Generates all the semantic slot colors based on the theme so far
+ * We'll use these as fallbacks for semantic slots that the passed in theme did not define.
+ * The caller must still mix in the customized semantic slots at the end.
+ */
+export function makeSemanticColors(
+  p: IPalette,
+  e: IEffects,
+  s: Partial<ISemanticColors> | undefined,
+  isInverted: boolean,
+  depComments: boolean,
+): ISemanticColors {
+  let toReturn: ISemanticColors = {
+    // DEFAULTS
+    bodyBackground: p.white,
+    bodyBackgroundHovered: p.neutralLighter,
+    bodyBackgroundChecked: p.neutralLight,
+    bodyStandoutBackground: p.neutralLighterAlt,
+    bodyFrameBackground: p.white,
+    bodyFrameDivider: p.neutralLight,
+    bodyText: p.neutralPrimary,
+    bodyTextChecked: p.black,
+    bodySubtext: p.neutralSecondary,
+    bodyDivider: p.neutralLight,
+    disabledBodyText: p.neutralTertiary,
+    disabledBodySubtext: p.neutralTertiaryAlt,
+    disabledBorder: p.neutralTertiaryAlt,
+    focusBorder: p.neutralSecondary,
+    cardStandoutBackground: p.white,
+    cardShadow: e.elevation4,
+    cardShadowHovered: '', // set in second pass
+    variantBorder: p.neutralLight,
+    variantBorderHovered: p.neutralTertiary,
+    defaultStateBackground: p.neutralLighterAlt,
+
+    // LINKS
+    actionLink: p.neutralPrimary,
+    actionLinkHovered: p.neutralDark,
+    link: p.themePrimary,
+    linkHovered: p.themeDarker,
+
+    // BUTTONS
+    buttonBackground: p.white,
+    buttonBackgroundChecked: p.neutralTertiaryAlt,
+    buttonBackgroundHovered: p.neutralLighter,
+    buttonBackgroundCheckedHovered: p.neutralLight,
+    buttonBackgroundPressed: p.neutralLight,
+    buttonBackgroundDisabled: p.neutralLighter,
+    buttonBorder: p.neutralSecondaryAlt,
+    buttonText: p.neutralPrimary,
+    buttonTextHovered: p.neutralDark,
+    buttonTextChecked: p.neutralDark,
+    buttonTextCheckedHovered: p.black,
+    buttonTextPressed: p.neutralDark,
+    buttonTextDisabled: p.neutralTertiary,
+    buttonBorderDisabled: p.neutralLighter,
+
+    primaryButtonBackground: p.themePrimary,
+    primaryButtonBackgroundHovered: p.themeDarkAlt,
+    primaryButtonBackgroundPressed: p.themeDark,
+    primaryButtonBackgroundDisabled: p.neutralLighter,
+    primaryButtonBorder: 'transparent',
+    primaryButtonText: p.white,
+    primaryButtonTextHovered: p.white,
+    primaryButtonTextPressed: p.white,
+    primaryButtonTextDisabled: p.neutralQuaternary,
+
+    accentButtonBackground: p.accent,
+    accentButtonText: p.white,
+
+    // INPUTS
+    inputBorder: p.neutralSecondary,
+    inputBorderHovered: p.neutralPrimary,
+    inputBackground: p.white,
+    inputBackgroundChecked: p.themePrimary,
+    inputBackgroundCheckedHovered: p.themeDark,
+    inputPlaceholderBackgroundChecked: p.themeLighter,
+    inputForegroundChecked: p.white,
+    inputIcon: p.themePrimary,
+    inputIconHovered: p.themeDark,
+    inputIconDisabled: p.neutralTertiary,
+    inputFocusBorderAlt: p.themePrimary,
+    smallInputBorder: p.neutralSecondary,
+    inputText: p.neutralPrimary,
+    inputTextHovered: p.neutralDark,
+    inputPlaceholderText: p.neutralSecondary,
+    disabledBackground: p.neutralLighter,
+    disabledText: p.neutralTertiary,
+    disabledSubtext: p.neutralQuaternary,
+
+    // LISTS
+    listBackground: p.white,
+    listText: p.neutralPrimary,
+    listItemBackgroundHovered: p.neutralLighter,
+    listItemBackgroundChecked: p.neutralLight,
+    listItemBackgroundCheckedHovered: p.neutralQuaternaryAlt,
+
+    listHeaderBackgroundHovered: p.neutralLighter,
+    listHeaderBackgroundPressed: p.neutralLight,
+
+    // MENUS
+    menuBackground: p.white,
+    menuDivider: p.neutralTertiaryAlt,
+    menuIcon: p.themePrimary,
+    menuHeader: p.themePrimary,
+    menuItemBackgroundHovered: p.neutralLighter,
+    menuItemBackgroundPressed: p.neutralLight,
+    menuItemText: p.neutralPrimary,
+    menuItemTextHovered: p.neutralDark,
+
+    errorText: !isInverted ? '#a4262c' : '#F1707B',
+
+    messageText: !isInverted ? '#323130' : '#F3F2F1',
+    messageLink: !isInverted ? '#005A9E' : '#6CB8F6',
+    messageLinkHovered: !isInverted ? '#004578' : '#82C7FF',
+
+    infoIcon: !isInverted ? '#605e5c' : '#C8C6C4',
+    errorIcon: !isInverted ? '#A80000' : '#F1707B',
+    blockingIcon: !isInverted ? '#FDE7E9' : '#442726',
+    warningIcon: !isInverted ? '#797775' : '#C8C6C4',
+    severeWarningIcon: !isInverted ? '#D83B01' : '#FCE100',
+    successIcon: !isInverted ? '#107C10' : '#92C353',
+
+    infoBackground: !isInverted ? '#f3f2f1' : '#323130',
+    errorBackground: !isInverted ? '#FDE7E9' : '#442726',
+    blockingBackground: !isInverted ? '#FDE7E9' : '#442726',
+    warningBackground: !isInverted ? '#FFF4CE' : '#433519',
+    severeWarningBackground: !isInverted ? '#FED9CC' : '#4F2A0F',
+    successBackground: !isInverted ? '#DFF6DD' : '#393D1B',
+
+    // Deprecated slots, later pass by _fixDeprecatedSlots() for self-referential slots
+    warningHighlight: !isInverted ? '#ffb900' : '#fff100',
+    warningText: '',
+    successText: !isInverted ? '#107C10' : '#92c353',
+    listTextColor: '',
+    menuItemBackgroundChecked: p.neutralLight,
+
+    // mix in customized semantic slots for second pass
+    ...s,
+  };
+
+  // second pass for self-referential slots
+  toReturn = {
+    ...toReturn,
+    cardShadowHovered: !isInverted ? e.elevation8 : '0 0 1px ' + toReturn.variantBorderHovered,
+  };
+
+  return _fixDeprecatedSlots(toReturn, depComments!);
+}
+
+function _fixDeprecatedSlots(s: ISemanticColors, depComments: boolean): ISemanticColors {
+  // Add @deprecated tag as comment if enabled
+  let dep = '';
+  if (depComments === true) {
+    dep = ' /* @deprecated */';
+  }
+
+  /* eslint-disable deprecation/deprecation */
+  s.listTextColor = s.listText + dep;
+  s.menuItemBackgroundChecked += dep;
+  s.warningHighlight += dep;
+  s.warningText = s.messageText + dep;
+  s.successText += dep;
+  /* eslint-enable deprecation/deprecation */
+  return s;
+}
